+++ CHANGELOG.md
@@ -3,3 +3,7 @@ Flutter Project Setup CLI Changelog
 ## 0.0.1
 
 - Implemented flutter_project_setup_cli create command template with specified args
+- Implemented dynamic links, and JWT authentication setups
+- Implemented State Management setup with Riverpod
+- Implemented Notification setup with Firebase Cloud Messaging
+- Implemented Firebase project integration